Force-framework: renamed tests packages.
