+++ src/test/java/br/com/schumaker/force/framework/reflection/ValueReflectionTest.java
@@ -1,9 +1,9 @@
-package br.com.schumaker.octopus.framework.reflection;
+package br.com.schumaker.force.framework.reflection;
 
-import br.com.schumaker.octopus.framework.annotations.bean.Value;
-import br.com.schumaker.octopus.framework.ioc.Environment;
-import br.com.schumaker.octopus.framework.exception.OctopusException;
-import br.com.schumaker.octopus.framework.ioc.reflection.ValueReflection;
+import br.com.schumaker.force.framework.annotations.bean.Value;
+import br.com.schumaker.force.framework.ioc.Environment;
+import br.com.schumaker.force.framework.exception.ForceException;
+import br.com.schumaker.force.framework.ioc.reflection.ValueReflection;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -137,6 +137,6 @@ public class ValueReflectionTest {
         environment.setKey("test.invalid.type", "invalid-Int");
 
         // Act & Assert
-        assertThrows(OctopusException.class, () -> valueReflection.injectParameterValue(parameter));
+        assertThrows(ForceException.class, () -> valueReflection.injectParameterValue(parameter));
     }
 }
